# services/api-worker/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "C360 API",
    "version": "0.1.0"
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Health check",
        "responses": {
          "200": { "description": "OK" }
        }
      }
    },
    "/tenants": {
      "get": {
        "summary": "List tenants",
        "parameters": [
          { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 1000 }, "description": "Max results (default 100)" },
          { "name": "offset", "in": "query", "schema": { "type": "integer", "minimum": 0 }, "description": "Offset for pagination (default 0)" }
        ],
        "responses": {
          "200": {"description": "List of tenants", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Tenant" } } } } },
          "400": {"description": "Bad Request", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},
          "500": {"description": "Server Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}}
        }
      },
      "post": {
        "summary": "Create tenant",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object", "properties": { "name": { "type": "string" } }, "required": ["name"] } } }
        },
        "responses": { "200": { "description": "Created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Tenant" } } } } }
      }
    },
    "/tenants/{tenantId}": {
      "get": {
        "summary": "Get tenant",
        "parameters": [ { "name": "tenantId", "in": "path", "required": true, "schema": { "type": "string" } } ],
  "responses": { "200": { "description": "Tenant", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Tenant" } } } }, "404": { "description": "Not found", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}} } }
      },
      "patch": {
        "summary": "Update tenant",
        "parameters": [ { "name": "tenantId", "in": "path", "required": true, "schema": { "type": "string" } } ],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "properties": { "name": { "type": "string" } } } } } },
  "responses": { "200": { "description": "Updated", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Tenant" } } } }, "400": { "description": "Bad Request", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}}, "404": { "description": "Not found", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}} } }
      },
      "delete": {
        "summary": "Delete tenant",
        "parameters": [ { "name": "tenantId", "in": "path", "required": true, "schema": { "type": "string" } } ],
  "responses": { "200": { "description": "Deleted" }, "404": { "description": "Not found", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}} } }
      }
    },
    "/tenants/{tenantId}/users": {
      "get": {
        "summary": "List users for a tenant",
        "parameters": [
          { "name": "tenantId", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 1000 }, "description": "Max results (default 200)" },
          { "name": "offset", "in": "query", "schema": { "type": "integer", "minimum": 0 }, "description": "Offset for pagination (default 0)" }
        ],
  "responses": { "200": { "description": "List of users", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/User" } } } } }, "400": { "description": "Bad Request", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}}, "404": { "description": "Not found", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}} }
      },
      "post": {
        "summary": "Create user",
        "parameters": [ { "name": "tenantId", "in": "path", "required": true, "schema": { "type": "string" } } ],
  "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "properties": { "email": { "type": "string", "format": "email" }, "role": { "type": "string", "enum": ["admin", "member"] } }, "required": ["email"] } } } },
  "responses": { "201": { "description": "Created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/User" } } } }, "400": { "description": "Bad Request", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}}, "404": { "description": "Not found", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}} }
      }
    },
    "/tenants/{tenantId}/users/{userId}": {
      "get": {
        "summary": "Get user",
        "parameters": [
          { "name": "tenantId", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "userId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
  "responses": { "200": { "description": "User", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/User" } } } }, "404": { "description": "Not found", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}} }
      },
      "patch": {
        "summary": "Update user",
        "parameters": [
          { "name": "tenantId", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "userId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
  "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "properties": { "email": { "type": "string", "format": "email" }, "role": { "type": "string", "enum": ["admin", "member"] } } } } } },
  "responses": { "200": { "description": "Updated", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/User" } } } }, "400": { "description": "Bad Request", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}}, "404": { "description": "Not found", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}} }
      },
      "delete": {
        "summary": "Delete user",
        "parameters": [
          { "name": "tenantId", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "userId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
  "responses": { "200": { "description": "Deleted" }, "404": { "description": "Not found", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}} }
      }
    }
  },
  "components": {
    "schemas": {
      "Tenant": {
        "type": "object",
        "properties": {
          "tenant_id": { "type": "string" },
          "name": { "type": "string" },
          "created_at": { "type": "string", "format": "date-time" }
        },
        "required": ["tenant_id", "name"]
      },
      "User": {
        "type": "object",
        "properties": {
          "user_id": { "type": "string" },
          "tenant_id": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "role": { "type": "string", "enum": ["admin", "member"] },
          "created_at": { "type": "string", "format": "date-time" }
        },
        "required": ["user_id", "tenant_id", "email", "role"]
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": { "type": "string" },
              "message": { "type": "string" }
            },
            "required": ["code", "message"]
          }
        },
        "required": ["error"]
      }
    }
  }
}